requirements.txt
python-dotenv==1.1.0
openai==1.77.0
redis>=4.5.0

camel-ai==0.2.57
typing-extensions>=4.5.0
asyncio>=3.4.3 
numpy>=1.22.0
tqdm>=4.66.0

pillow>=10.0.0
pyyaml>=6.0.0
requests_oauthlib>=1.3.0
sqlalchemy>=2.0.0
pandas>=1.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
graphiti-core==0.3.21
langchain-openai==0.3.16
langgraph==0.4.3
ipywidgets==8.1.7
pytest==8.3.5

flask==2.0.1
flask-cors==5.0.1
Flask-SocketIO==5.5.0
neo4j==5.28.1
python-socketio==5.4.0
python-engineio==4.2.1
langchain==0.3.25
langchain-community==0.3.23

# supabase==1.0.3

qdrant-client>=1.7.0
pymilvus>=2.3.0

requests
distro
jsonschema
pyyaml
pydantic_settings
starlette
oauthlib
pandas
tiktoken

joblib

rag_helper
